verification refactoring
