--- verification/testdata/truststore/x509/trust-store-with-cert-symlinks/GlobalSignRootCA_SYMLINK.crt
@@ -1,1 +1,0 @@
-GlobalSignRootCA.crt
