+++ verifier/helpers_test.go
@@ -3,7 +3,6 @@ package verifier
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -11,7 +10,7 @@ import (
 
 	"github.com/notaryproject/notation-core-go/signature"
 	"github.com/notaryproject/notation-go/dir"
-	"github.com/notaryproject/notation-go/verification/trustpolicy"
+	"github.com/notaryproject/notation-go/verifier/trustpolicy"
 )
 
 func dummyPolicyStatement() (policyStatement trustpolicy.TrustPolicy) {
@@ -116,49 +115,5 @@ func TestLoadX509TrustStore(t *testing.T) {
 	}
 	if len(caCerts) != 3 || len(saCerts) != 3 {
 		t.Fatalf("Both of the named stores should have 3 certs")
-	}
-}
-
-// TestApplicableTrustPolicy tests filtering policies against registry scopes
-func TestApplicableTrustPolicy(t *testing.T) {
-	policyDoc := dummyPolicyDocument()
-
-	policyStatement := dummyPolicyStatement()
-	policyStatement.Name = "test-statement-name-1"
-	registryScope := "registry.wabbit-networks.io/software/unsigned/net-utils"
-	registryUri := fmt.Sprintf("%s@sha256:hash", registryScope)
-	policyStatement.RegistryScopes = []string{registryScope}
-	policyStatement.SignatureVerification = trustpolicy.SignatureVerification{VerificationLevel: "strict"}
-
-	policyDoc.TrustPolicies = []trustpolicy.TrustPolicy{
-		policyStatement,
-	}
-	// existing Registry Scope
-	policy, err := getApplicableTrustPolicy(&policyDoc, registryUri)
-	if policy.Name != policyStatement.Name || err != nil {
-		t.Fatalf("getApplicableTrustPolicy should return %q for registry scope %q", policyStatement.Name, registryScope)
-	}
-
-	// non-existing Registry Scope
-	policy, err = getApplicableTrustPolicy(&policyDoc, "non.existing.scope/repo@sha256:hash")
-	if policy != nil || err == nil || err.Error() != "artifact \"non.existing.scope/repo@sha256:hash\" has no applicable trust policy" {
-		t.Fatalf("getApplicableTrustPolicy should return nil for non existing registry scope")
-	}
-
-	// wildcard registry scope
-	wildcardStatement := dummyPolicyStatement()
-	wildcardStatement.Name = "test-statement-name-2"
-	wildcardStatement.RegistryScopes = []string{"*"}
-	wildcardStatement.TrustStores = []string{}
-	wildcardStatement.TrustedIdentities = []string{}
-	wildcardStatement.SignatureVerification = trustpolicy.SignatureVerification{VerificationLevel: "skip"}
-
-	policyDoc.TrustPolicies = []trustpolicy.TrustPolicy{
-		policyStatement,
-		wildcardStatement,
-	}
-	policy, err = getApplicableTrustPolicy(&policyDoc, "some.registry.that/has.no.policy@sha256:hash")
-	if policy.Name != wildcardStatement.Name || err != nil {
-		t.Fatalf("getApplicableTrustPolicy should return wildcard policy for registry scope \"some.registry.that/has.no.policy\"")
 	}
 }
