+++ verifier/testdata/truststore/x509/ca/trust-store-with-invalid-certs/invalid
