+++ verifier/testdata/truststore/x509/ca/trust-store-with-leaf-certs/GlobalSignRootCA.crt
