+++ verifier/testdata/truststore/x509/ca/valid-trust-store-2/GlobalSignRootCA.crt
