+++ verifier/testdata/truststore/x509/ca/valid-trust-store-self-signed/openssl-minimum-self-signed.pem
