+++ verifier/testdata/truststore/x509/ca/valid-trust-store/GlobalSign.der
