+++ verifier/testdata/truststore/x509/ca/valid-trust-store_SYMLINK/valid-trust-store_SYMLINK
