+++ verifier/testdata/truststore/x509/signingAuthority/valid-trust-store-2/GlobalSign.der
