+++ verifier/testdata/truststore/x509/signingAuthority/valid-trust-store/GlobalSign.der
