+++ verifier/testdata/truststore/x509/trust-store-with-cert-symlinks/GlobalSignRootCA.crt
