+++ verifier/testdata/truststore/x509/trust-store-with-cert-symlinks/GlobalSignRootCA_SYMLINK.crt
@@ -1,0 +1,1 @@
+GlobalSignRootCA.crt
