+++ verifier/testdata/truststore/x509/trust-store-with-directories/GlobalSignRootCA.crt
