+++ verifier/trustpolicy/trustpolicy_test.go
