+++ verifier/truststore/truststore_test.go
